+++ NEWS.md
@@ -1,15 +1,15 @@
 
 # Release updates
 
-<!-- 
+
 ## alced.api version 1.1.3
-* Added functionality to do immediate geographic/temporal aggregation
 * Resized README example data retrieval
 * Updated region names and codes based on new ACLED API guide
 * Bug fix in region call
+* Added support function module
 * 1 merge request by Rob Williams https://github.com/jayrobwilliams 
 + !3 Bug fix: %20 instead of % for white spaces in country names
- -->
+
 
 ## alced.api version 1.1.2
 * 1 merge request by Rob Williams https://github.com/jayrobwilliams